fixed some issue and cleanup code
