--- src/app/top-bar/top-bar.component.html
@@ -1,62 +1,0 @@
-<nav class="navbar navbar-expand-md navbar-dark fixed-top bg-top-color">
-  <button
-    class="navbar-toggler"
-    type="button"
-    style="outline: none;"
-    (click)="toggleSidenav()"
-  >
-    <div class="navbar navbar-dark">
-      <i class="fas fa-bars"></i>
-    </div>
-  </button>
-
-  <div class="w-100">
-    <div class="dropdown float-right">
-      <div class="dropbtn pt-2">{{ getUserFirstLetter() }}</div>
-
-      <div class="dropdown-content"></div>
-      <i
-        class="fas fa-sign-out-alt cursor-pointer logoutDesign"
-        data-target="#myModal1"
-        data-toggle="modal"
-        ><a class=""></a
-      ></i>
-    </div>
-    <div class="userName float-right my-4 mx-0">Hi, {{ getUserName() }}</div>
-  </div>
-</nav>
-<!-- logout modal -->
-<div class="modal fade" id="myModal1">
-  <div class="modal-dialog modal-dialog-centered popup">
-    <div class="modal-content p-4 ">
-      <div class="modal-header border-0 pt-0">
-        <div class="modal-title">
-          <h4>Are you sure you want to Logout?</h4>
-        </div>
-      </div>
-
-      <div class="modal-body">
-        <div class="col-md-6 float-right">
-          <button
-            type="submit"
-            (click)="logout()"
-            data-dismiss="modal"
-            class="btn btn-yes"
-          >
-            Logout
-          </button>
-        </div>
-
-        <div class="col-md-6 float-left ">
-          <button
-            type="button"
-            class="btn btn-cancel float-right"
-            data-dismiss="modal"
-          >
-            Cancel
-          </button>
-        </div>
-      </div>
-    </div>
-  </div>
-</div>
