--- src/app/top-bar/top-bar.component.scss
@@ -1,89 +1,0 @@
-// component css goes here
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-  font-family: arial;
-}
-textarea {
-  border: none;
-  outline: none;
-  width: 100%;
-  min-height: 50px;
-  resize: none;
-}
-textarea:focus {
-  border: none;
-}
-
-.loginCont {
-  width: 400px;
-  height: 450px;
-  padding: 20px;
-  background: #fff;
-  box-shadow: 0 2px 5px rgba(0, 0, 0, 0.2);
-  position: absolute;
-  top: 0;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  margin: auto;
-}
-input {
-  width: 100%;
-  border: none;
-}
-input:focus {
-  outline: none;
-  border: none;
-  background: #fff;
-}
-.inputCont {
-  font-size: 14px;
-  width: 100%;
-  border-radius: 3px;
-  border: 1px solid grey;
-  line-height: 30px;
-  margin-bottom: 10px;
-  padding-left: 5px;
-}
-
-.fLeft {
-  float: left;
-}
-.fRight {
-  float: right;
-}
-.forgotTxt {
-  font-size: 12px;
-}
-.otpBtn {
-  width: 100%;
-  height: 30px;
-  margin-bottom: 5px;
-  cursor: pointer;
-  background: #a23f9d;
-  color: #fff;
-  border: none !important;
-  font-size: 14px;
-  font-weight: bold;
-}
-.cursor {
-  cursor: pointer;
-}
-.disable {
-  opacity: 0.5;
-}
-.enable {
-  opacity: 1 !important;
-}
-.fs14 {
-  font-size: 14px;
-}
-.bg-gray {
-  // background-color: #e5e5e5; //old
-  background-color: #f3f4f4; //new
-  min-height: 100vh;
-  height: 100%;
-  display: flow-root;
-}
