+++ src/app/user-details/user-details.component.scss
@@ -23,8 +23,8 @@
   transition: ease 1s;
 }
 .photoSize {
-  height: 50%;
-  width: 50%;
+  height: 150px;
+  width: 120px;
 }
 .zoomImg {
   width: 100%;
