+++ src/app/users/users.component.html
@@ -1,4 +1,57 @@
 <div class="header">
+  <div class="w-100">
+    <div class="userDetail">
+      <div class="dropdown float-right">
+        <div class="dropbtn pt-2">{{ getUserFirstLetter() }}</div>
+
+        <div class="dropdown-content"></div>
+        <i
+          class="fas fa-sign-out-alt cursor-pointer logoutDesign"
+          data-target="#myModal1"
+          data-toggle="modal"
+          ><a class=""></a
+        ></i>
+      </div>
+      <div class="userName float-right my-4 mx-0">Hi, {{ getUserName() }}</div>
+
+      <!-- </nav> -->
+      <!-- logout modal -->
+      <div class="modal fade" id="myModal1">
+        <div class="modal-dialog modal-dialog-centered popup">
+          <div class="modal-content p-4 ">
+            <div class="modal-header border-0 pt-0">
+              <div class="modal-title">
+                <h4>Are you sure you want to Logout?</h4>
+              </div>
+            </div>
+
+            <div class="modal-body">
+              <div class="col-md-6 float-right">
+                <button
+                  type="submit"
+                  (click)="logout()"
+                  data-dismiss="modal"
+                  class="btn btn-yes"
+                >
+                  Logout
+                </button>
+              </div>
+
+              <div class="col-md-6 float-left ">
+                <button
+                  type="button"
+                  class="btn btn-cancel float-right"
+                  data-dismiss="modal"
+                >
+                  Cancel
+                </button>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+  </div>
   <div class="pt-2 pb-2">
     <span class="float-left page-header">All Users</span>
     <span class="smallPageHeader" *ngIf="users">({{ users.length }})</span>
@@ -9,9 +62,7 @@
       class="btn btn-color float-right"
       data-toggle="modal"
       *ngIf="
-        userRoles &&
-        userRoles['userManagement'] &&
-        userRoles.userManagement[0].Add
+        userRoles && userRoles.userManagement && userRoles.userManagement[0].Add
       "
     >
       <i class="fas fa-plus"></i>
@@ -38,7 +89,7 @@
       <div class="float-right mr-2 pl-2">
         <select
           class="userDropDown fs14"
-          (change)="filterUsers()"
+          (click)="filterUsers()"
           [(ngModel)]="filterUserValue"
         >
           <option class="fs14" [ngValue]="null">All Users</option>
@@ -80,7 +131,7 @@
               scope="col"
               *ngIf="
                 userRoles &&
-                userRoles['userManagement'] &&
+                userRoles.userManagement &&
                 userRoles.userManagement[0].Delete
               "
             >
@@ -119,7 +170,7 @@
               (click)="deletStoreId(user.id)"
               *ngIf="
                 userRoles &&
-                userRoles['userManagement'] &&
+                userRoles.userManagement &&
                 userRoles.userManagement[0].Delete
               "
             >
@@ -164,9 +215,7 @@
     <div
       class="container"
       *ngIf="
-        userRoles &&
-        userRoles['userManagement'] &&
-        userRoles.userManagement[0].Add
+        userRoles && userRoles.userManagement && userRoles.userManagement[0].Add
       "
     >
       <div class="modal fade" id="userModal">
@@ -385,7 +434,7 @@
       class="container"
       *ngIf="
         userRoles &&
-        userRoles['userManagement'] &&
+        userRoles.userManagement &&
         userRoles.userManagement[0].Delete
       "
     >
