+++ src/app/users/users.component.scss
@@ -88,7 +88,7 @@ input[type="number"]::-webkit-outer-spin-button {
 }
 .table td,
 th {
-  max-width: 200px;
+  max-width: 100%;
   text-overflow: ellipsis;
   overflow: hidden;
 }
@@ -118,8 +118,7 @@ th {
 
 .list-table-scroll-tbody {
   display: block;
-  height: 100%;
-  width: 100%;
+  height: 420px;
   overflow: auto;
 }
 .table-fixed-head thead {
@@ -131,4 +130,85 @@ th {
   display: table;
   width: 100%;
   table-layout: inherit;
+}
+.userDetail {
+  display: inline-flexbox;
+  width: 100%;
+  align-self: flex-end;
+}
+
+.dropbtn {
+  border-radius: 100px;
+  background-color: #c4c4c4;
+  display: inline-block;
+  width: 47px;
+  height: 47px;
+
+  font-family: Noto Sans;
+  font-style: normal;
+  font-weight: normal;
+  font-size: 22px;
+  line-height: 30px;
+  text-align: center;
+}
+.productLabel {
+  font-size: 15px;
+  color: #818181;
+}
+.productValue {
+  font-size: 13px;
+  color: #000000;
+}
+.dropbtn {
+  border-radius: 100px;
+  background-color: #c4c4c4;
+  display: inline-block;
+  width: 47px;
+  height: 47px;
+
+  font-family: Noto Sans;
+  font-style: normal;
+  font-weight: normal;
+  font-size: 22px;
+  line-height: 30px;
+  text-align: center;
+}
+.btn-color {
+  border: none;
+  border-radius: 3px;
+  background-color: #ff1a3d;
+  color: #ffffff;
+  width: 20%;
+  font-weight: 600;
+}
+.cardBox {
+  box-shadow: 0px 2px 6px 2px #00000047;
+}
+.cardBox:hover {
+  box-shadow: 0px 1px 3px 1px #00000047;
+  transition: ease 1s;
+}
+.photoSize {
+  height: 150px;
+  width: 120px;
+}
+.zoomImg {
+  width: 100%;
+  height: 100%;
+}
+.close {
+  outline: none;
+}
+.popup {
+  max-width: 1000px !important;
+}
+.dropdown-content {
+  display: none;
+  position: fixed;
+  background-color: #f9f9f9;
+  min-width: 200px;
+  box-shadow: 0px 8px 16px 0px rgba(0, 0, 0, 0.2);
+  z-index: 1;
+  right: 1rem;
+  border-radius: 5px;
 }
