+++ src/assets/css/bootstrap.css
@@ -7731,7 +7731,7 @@ button.bg-dark:focus {
 }
 
 .w-100 {
-  width: 80% !important;
+  width: 100% !important;
 }
 
 .w-auto {
